DVID stack wizard: create stack group for ortho-views
If DVID based ortho-views are added and a new stack is created, a stack
group will now be created automatically. Its name is the base name of
the added stacks. This makes opening the created ortho-views at once in
the front-end possible.

This is currently only done if a new project is created, because it
avoids asking the user which project a new stack group should be created
in.

See catmaid/CATMAID#1081