Window maker: add default selection to select element creation
The "use numbers" options is replaced by a default selection value. The
number option wasn't used and since it is better to be explicit than
implicit, the only way to use other select element values than a
non-object itself, is to define it through the value property of an object.

Also add support for change handlers and titles for select elements.
Additionally, a new object item handler has been added for appending
controls to a tab.