3D viewer: add explicit visibility order mode for animations
This mode allows to define a pattern to explicitly state which skeleton
should become visible after how many rotations. A pattern looks like
this: "i(0: id1); (2: id2, id3); (3: id4)". In this case skeleton with
id1 would be visible right from the start, skeletons with ids id2 and
id3 would be visible after the second rotation and skeleton with id4
would be visible after rotation 3.

The dialog isn't very user friendly yet, because it doesn't check the
entered pattern and it doesn't allow to edit a previous pattern.

See catmaid/CATMAID#1207