Add button panel toggle button to many widgets
Many tracing widgets have a lot of controls and would benefit from being
able to hide them once they show the data one is interested in.
Therefore, this commit adds a little icon to the title bar of many
tracing widgets which allows to toggle the display of these buttons.

This is also added to the createWidget() support function so that
widgets using this don't need to explicitly add this functionality if
they have a button panel.