Authentication: add any project role decorator
Add a decorator that checks if the user has a given role (or is an
administrator) for any project, not just for a particular project
associated with the request. This is useful for access control for
actions that are not associated with a particular project, but should
still have some sane limits to access.