Create backend storage for client data
Remove the unused backend Settings models and replace it with two
models, ClientDatastore and ClientData, that provide a generic, scoped
key-value store for client data.

ClientDatastore defines groups of client data (e.g., settings) that
exist independently.

ClientData provides a key-value store for JSON-encoded client data; the
backend is agnostic to the content of client data other than that it is
valid JSON. In addition to a datastore, data key-value entries are
associated with an optional project and user. This provides four scopes
of entries for each datastore and key pair: global (no project, no
user), project-default (project but no user), user-default (user but no
project), and user-project specific (both user and project). When
retrieving key-value entries, as many of these four scopes that are
specified will be returned, so that the client can be responsible for
reconciling them in a manner semantically meaningful for the data
domain.

This model and API allow the backend to enforce access control and
provide persistence to client data, while keeping the client responsible
for the content, correctness and semantics of its data.

See catmaid/CATMAID#965.