Skeleton sources: add basic subscription infrastructure
This makes skeleton sources able of registering to other skeleton
sources. A subscription can parameterized to indicate color syncing a
combinatorial operation with which it will be combined with other
subscriptions and a filter. The filter can be used to only listen to
either additions, removals or updates in a source. A button-panel like
UI to edit and display subscriptions can be added to widgets so that it
is accessible through a title bar button.

See catmaid/CATMAID#1000