Source subsctions: move more functionality into subscription type
A subscription requires now to register with a target before it actually listens
to events. It now expects events to supply models as object with IDs as field
names. It is now possible and set as default to only consider selections as
actual append action. This is the same behavior that is currently in use with
regular skeleton source appending.