DOM: make toggle button functions accept optional title argument
This allows to display some hover information on them. Widget
configuration can now have an optional subscriptionSource parameter. If
it is falsy or not used, the widget itself is expected to be the source.
Alternatively, an array can be passed in where the first element is the
source and the second one is the title.