Move treenode-specific APIs from node to treenode
Move these endpoints and their routes from the node controller to the
treenode controller. The confidence update API could apply to nodes in
general, because connector nodes have a confidence field, but as written
it only functions for treenodes, so it is move also.

Update the routes to match new conventions, although they are not yet
documented endpoints.