Graph widget: fix path finding with groups
The internal source and target sets for path finding consisted of the
actual skeleton IDs and not node IDs, how it is expected by the
collector function. Using the collector function makes sense so that
groups can be evaluated dynamically and so this commit makes the source
and target sets consist of node IDs instead of already expanded
skeleton IDs.