Color picker: try to fix position only after set-up is complete
Unfortunately, the color picker library we use will create the actual DOM
elements only after it has been added to the DOM and is shown. A user doesn't
notice, but so far one of our callback functions expected the color picker to
already of its final display size when it is shown. This is now changed so that
this callback prepares a resize function which is then called after as soon as
the final layout is set. This has to be done over two different callbacks,
because onShow() has the required color picker nodes and we utilize the first
onRender() call after onShow() has been called to get the timing right.

Fixes catmaid/CATMAID#1238