Upgrade Django from v1.6.11 to v1.7.11
Django 1.8 was released beginning of April 2015. This caused Django 1.6
to become unsupported. To be able to receive further security (and
general) updated, this upgrade is performed.

Django 1.7 comes with many changes and some of them require action on
our side. A detailed view on those can be found here:

  https://docs.djangoproject.com/en/1.7/releases/1.7/

The biggest change is that Django now comes with its own migration
system. Therefore, South is removed and a new migration history is
started. The complete existing SQL schema is created by the initial
migration of the new system. This migration will be ignored in regular
updates if all previous South migrations have been applied. If a regular
migration is tried from a non-empty database state not matching the last
South migration (0061_remove_unused_api_key_table), an error is thrown
and the user is told to bring it to this state (along with a rough
guideline how).

To updating an existing instance to this version, please make sure the
database was migrated to the last South migration first (release
2015.12.21). Then, with the new CATMAID version fake the initial
migrations (the tables are already there):

   manage.py migrate catmaid --fake 0001_initial

Afterwards, you can just 'manage.py migrate' as you would normally do.

This commit fixes all the required things to be able to update in the
first place. For CATMAID, this is in particular:

- Start a new migration history with Django's new migration system for both
  catmaid and performancetests Django applications
- Use AppConfig.ready() for back-end initialization
- Change post_syncdb signal to post_migrate
- Upgrade django-taggit from v0.11.2 to v0.14.0 for Django 1.7 support
- Prevent automatic auto-discover for admin app. This is required for
  django-adminplus to work with Django 1.7.
- Upgrade jsonfield from v1.0.0 to v1.0.3 to include a south migration
  fix.
- Upgrade kombu from v3.0.12 to v3.0.32 to remove deprecations
- Upgrade django-celery from v3.1.9 to v3.1.17 to remove deprecations
- Make sure TEMPLATE_DIRS is a tuple.
- Update use of create_superuser management command.
- Replace SortedDict with OrderedDict
- Replaced last uses of mimetype in responses with content_type
- Add missing fields attribute for ModelForm classes
- Update custom database engine so that no override happens for in-memory DB
  connections.