Update Django from v1.7.11 to v1.8.7
This is another major upgrade of Django. Version 1.8 is a long term
support release and will receive security updates at least until April
2018. Besides security enhancements, Django 1.8 brings introduces some
features that might be useful to CATMAID. E.g. new Postgres specific
functionality [1].

Some dependencies had to be upgraded to work with Django 1.8 and minor
changes in CATMAID were required:

- Django-Guardian has been updated from v1.2.5 to v1.3.1. The more recent
  version 1.3.2 has a problem with PostgreSQL at the moment.
- Django's form wizards (used for e.g. DVID import) are now part of a
  stand-alone package: django-formtools. It has been included as
  dependency.
- Upgrade django-adminplus from v0.2 to v0.5.
- The returned object of _meta.get_all_related_objects() changed.
- The database backend API changed slightly, our custom psycopg wrapper
  had to be updated.
- Update management commands to use argparse argument definitions.
- Templates loaded get_template() or select_template don't expect a Context
  instance anymore in their rendeR() method. They want a dict instead.

I went through the list of breaking and non-breaking changes and changes
not included in this commit shouldn't be required this time.

[1] https://docs.djangoproject.com/en/1.9/ref/contrib/postgres/