Upgrade django-pipeline from v1.3.25 to v1.6
This is done in preparation of upgrading to Django 1.8. Since there seems to be
a problem with using the CSS compressor we used before, it is disabled for now.
Also only the app dir static file finder is now used. This means that static
files are only looked for in the static/ directories of all installed
applications. This should be okay with our current setup, but it needs to be
updated if there should be more sources for static files.