Tracing overlay: color tracing by skeleton source
Add a skeleton source to the tracing overlay which can subscribe to
other skeleton sources to determine display options for skeletons in the
overlay, including color and visibility. Skeletons present in this
source will use the skeleton model color instead of the pre-configured
colors.

Create a tracing overlay setting in the settings widget to also hide
skeletons that are not present in the display skeleton source.

Closes catmaid/CATMAID#895.