Volumes: work around GeoDjango problem with parsing PostGIS geometry
Since Django 1.8 Django fails to parse our EWKB geometry data. It fails
with "GEOS_ERROR: ParseException: Unknown WKB type 15". So apparently
GEOS doesn't recognize the type of our geometries, which are currently
stored as polyhedral surfaces. This commit can only work around this
problem by not having Django parse the geometry field when a volume
collection is requested. This in turn is done by not queriing it in the
first place, luckily it is not expected to be retuned from this endpoint
anyway.

Fixes catmaid/CATMAID#1250