Use Django's own JSONField and jsonb in Postgres
This also changes the type of the database columns that used JSON data
before, represented as strings. Unfortunately, we can't remove the
jsonfield dependency just yet, CATMAID's initial migration still
references models of this app, so unless we squash both the newer and
the initial migration, we have to keep this dependency.