Graph widget: fix removal of split neuron nodes
When removing a node that is part of a split neuron subgraph using the
"Remove" button in the "Selection" tab, also remove all other nodes in
the split subgraph. Also fix a related error when removing all split
nodes then adding a new neuron. Fix this same error in the case of
removing a split neuron through the removeSkeletons methods.

Closes catmaid/CATMAID#1253.