Skeleton sources: confirm when loading many models
When loading a source with many models, prompt for confirmation if the
number of models is greater than some threshold.

For now, this threshold is set at 1000 models for the selection table
and 3D viewer, and 50 for all other sources. In the future these values
could be configurable settings.

Also fix showing an information message if attempting to load an empty
source.

Closes catmaid/CATMAID#1240.