Color picker: fix color update of color picker enabled controls
CATMAID's color picker adapter overrode the original color picker's
renderCallback option. This was required to be able properly resize the
color picker when it was first rendered (introduced with d8b8f0e8).
However, this disabled also the color update of color picker enabled
elements. With this commit the displayCallback option is used, which is
called from the original color picker's renderCallback implementation.
This way we can have both our resize behavior and DOM element coloring.