Move generic CATMAID.fetch() method into catmaid-lib
The new method doesn't depend on jsonResponseHandler() anymore. This
causes no problems, because the response handler was explicitly marked
as silent anyway. In case of failure the returned promise was also
rejected (because reject() was passed in as error handler before). This
is the same what the new function does.