Selection table: apply sorting after name service update
Fixes a bug from d6fc60ef0adc84ccdae1017d523adf6a97672ee2 where an
exception is thrown when (a) names include data that must be retrieved
from the backend, e.g., annotations, (b) skeletons are appended to
the selection table that are not yet managed by the name service, and
(c) the table is sorted by name. The exception is thrown because the
name service returns null for the names while attempting to sort.

Instead perform the sorting after the name service is updated but before
the selection table GUI refresh.

Thanks to Katharina Eichler for the report and debug info.