Volumes: add general triangle mesh back-end
This TIN based PostGIS volume can be created through the API. To do so,
the "trimesh" type has to be used and a mesh parameter has to be
included. This mesh parameter has to be a JSON serialized list of two
lists: [[points], [triangles]]. The list of points contains lists of
three numbers, each one representing a vertex in the mesh. The array of
triangles also contains three element lists as items. Each one
represents a triangle based on the points in the other array, that are
referenced by the triangle index values.