Convex hull volumes: Add basic rule based hull generation
Node selection strategies can be used to constrain the nodes used from
skeletons in a selected source. These strategies are wrapped in filter
rules which allow optional filter application constraints like skeleton
ID and skeleton name. For now the most simple rule and strategy is
implemented: take all nodes of all skeletons. After nodes have been
filtered, a convex hull is created around them. This hull can be stored
in the back-end.

With the help of two functions of the new ConvexHullVolume constructor,
it is easy to create new meshes and show them in a currently open 3D
viewer:

  var rules = [];
  var skeletons = SkeletonAnnotations.activeSkeleton.getSelectedSkeletonModels();

  // Show in 3D viewer
  CATMAID.ConvexHullVolume.showCompartment(skeletons, rules, function(p) {
    console.log("Done"); console.log(p);
  });

  // Mesh generation only
  CATMAID.ConvexHullVolume.createTriangleMesh(skeletons, rules, function(p) {
    console.log("Done"); console.log(p);
  });

With the help of the rule list it is possible to specify more detailed
what meshes should be generated. It allows for various filter rules.
Also, for both functions referenced above, there is exists a plural
version (showCompartments, createTriangleMeshes) which is more efficient
when many different meshes should bes generated.