Convex hull volumes: support merge mode selection for rules
A merge mode is a set operation, either union or intersection. Node
contributions from individual rule/skeleton combinations are combined
with the existing set in a strict left-associative fashion. This should
be expressive enough for most use cases.

Also add a constants source file to include some constants that are used
in different contexts.