DOM: add custom content select element creation function
A custom content select is a regular select element which shows a custom
DIV container when clicked. This container can contain anything, a
checkbox select creation function is added as example. Such a checkbox
select looks very similar to a regular select element, but in addition
provides checkboxes in front of each element.