Connector table: fix 'show all' selection in table
If all connectors should be shown, DataTables will send -1 as display
length. The CATMAID back-end somehow expected 0 and used this as a
default if the parameter wasn't specified at all and to test if all
connectors should be returned. This is no unified to the value of -1.

Fixes catmaid/CATMAID#1271