Move geometry tools into own namespace
The convex hull and intersection source files are maintained in a folder
separate from all other CATMAID library code. This makes sense, because
it is actually another library for the most part and this is also why it
should live it in its own namespace, independent from when all other
CATMAID source files are loaded. Before it was added to the CATMAID
namespace and loading order was important. Using a separate namespace
makes loading of this dependency easier.