geometry.js: add alphaShape; see:
https://github.com/mikolalysenko/alpha-shape

It's similar to a convex hull but not necessarily convex,
allowing for a tighter fit. The extra paramater, the alpha,
is related to which tetragons to chip away from the convex hull.
First the delaunay is computated for all provided points, which
are then filtered by whether the circumradius of the tetragon
times the alpha is smaller than 1--which is a way of filtereing
for tetragons that are too large.