Add fall-back overview image (in tile layer)
If an image stack doesn't provide an overview image, a grid is now
displayed. The width and height of this artificial overview image is
following the ratio of the original stack. All tile sources that have
been using the dummy overview layer (which doesn't display anything) now
use the new artificial overview layer.

Closes catmaid/CATMAID#1259