Add utility method to find list in request dictionary
Traditionally, the CATMAID front-end sends lists form-encoded, i.e. the
list a = [1,2,3] would be sent as a[0]=1&a[1]=2&a[2]=3. Other
frameworks, e.g. jQuery, will send the same list as a=1&a=2&a3. The
function added is able to understand both formats and gives precedence
to the traditional CATMAID representation.