Make CATMAID.fetch() reject failed requests properly
Relying on automatic rejection of the CATMAID.fetch() promise doesn't
work in this case. An exception is thrown during response validation if
the response is invalid. However, since validation happens in an async
callback, such an exception isn't thrown in the scope of the enclosing
promise and the promise isn't rejected automatically. Therefore, a
try/catch block is added to explicitly reject the CATMAID.fetch()
promise.