Upgrade pixi.js to v4-dev
The current development version of Pixi.js seems stable for our use.
Tile browsing works without problems and blend modes can be assigned to
filters. This finally adds support for colored in-browser composite.
Unfortunately, there is still a bug in Pixi.js, which makes this only
work when a no-op filter is added to each layer in use.

See catmaid/CATMAID#1282