Add a command store to map changed IDs during redo
If commands create new database objects (like connectors, nodes,
neurons, etc.) and are undone and redone, the ID of such objects
changes. To keep track of these changes, commands store a mapping
between these IDs. If commands executed later references an ID that
was changed during redo of the earlier command, they can ask the for a
mapped ID. Likewise they are expected to push new IDs they created to
the mapper.