Re-optimize update behavior on connector and node creation
Until the last event change, updating the tracing layer was not done
before a node/connector was created manually, this way a potentially
expensive back-end call could be avoided. This wasn't working anymore
since a model node/connector creation event also triggers an update,
because this event is happening before the continuation of the
node/connector creation command. With this commit, this behavior is
restored by suspending the layer temporarily.