Add local node state checking and test before removing a node
When removing a node, a state is now required to be passed to the
back-end. This state consists of the node's edition time, its parent ID
plus edition time as well as its children IDs plus edition times.
Additionally, links to connectors are checked.

This local node state check or its components can be re-used easily by
other endpoints that want to make sure a client knows about the current
state.

If the expected state is matched, a row lock is acquired for the node.
This is required as a safeguard for undo/redo operations after long
browser sessions.