Node back-end: return generic representation of connectors
This makes state generation easier and makes the /node/list endpoint
support generic connectors (and changes its response format). Each link
between a treenode and a connector now explicitly refers to its
relation. In addition to that, a dictionary mapping relation IDs to
relation names is returned as an additional top level entry.