Window manager: add tabbed node
Add a tabbed window node, so that windows dragged onto each other can
occupy the same region in the window BSP. Each window in a tabbed group
is activatable by clicking on a tab.

Note that in this implementation, tabbed window nodes themselves can not
be dragged. However, the active tab window can be dragged as normal.

Closes catmaid/CATMAID#1021.