Classification: improve ontology instantiation tests dramatically
Now instead of testing each class_instance_class_instance link for an
ontology in its own query, all queries for one graphs and all features
are now part of one small recursive CTE. This is easier to read and much
faster. An additional improvement would be to combine all CTEs into one
query.

This is very useful for clustering.