Data views: lazy load images
For large collections of stacks and projects this reduces front page loading
time and wasted band-width significantly. Blazy [1] is used for the actual lazy
loading. It will monitor scrolling within the DIV container with ID "content"
and will load images in the current view that have the class "lazy" assigned.

[1] http://dinbror.dk/blazy/