Ensure settings client datastore exists
Because the migration creating the settings client datastore does not
exist in the new Django migrations system, make sure this datastore
exists for new installs. This could be accomplished by the client
without a migration, but would be much more complex and brittle to
future changes in, e.g., permissions.

Closes catmaid/CATMAID#1289.