State checks: add new front-end state representation
Besides different state components (children, parent, links, ...) there
are now also different state types. In a future commit, commands and
model functions will use instances of these types to obtain the view of
the state they need.  There is generic state which is extended by the
tracing layer. There is a local state, which is used by commands to let
them change the state based on undo and redo actions. And there is also
the so called no-check state, which indicates to the back-end that state
checks should be skipped.