Undo: always use original state on first command execution
The id/timestamp mapper will now also return the passed in values if the
command is executed for the first time. This increases robustness,
because the tracing layer and other state providers don't need to rely
on the mapping for its first execution (which is typically the most
important one).