Fix annotation cache update
Making the annotation cache use the jsonResponseHandler() function makes perfect
sense. However, when introduced in 13fc284, the execution context for the
success handler was accidentally changed to the global object. This broke the
annotation update, but is fixed now. The expected execution context for the
handler is the annotation cache.