Skeleton annotations: try to select parent of deleted nodes
The skeleton annotation object now listens to node deletion events. If a
removed node is the current active node, the active node is replaced. If
a parent ID is supplied, the parent is selected. Otherwise, the active
node will just become deselected.

This doesn't conflict with existing manual node removal handlers, they
deselect the active node in advance and reset it after the new handler
runs. But even if this handler would update the active node, there would
be no problem. In fact, this handler was added to streamline undo/redo
operations. Creating a node, undoing it and clicking again should result
in a new node being appended to what was selected before. This is
typically the parent, which is therefore selected by the new handler.
While not all cases are caught this way it mitigates the active node
selection problem after undo/redo operations.