Classification: fix classification root retrieval
The just recently added fix to return an empty set if classes or relations are
missing, didn't work as expected. Since a class map and a relation map can be
passed in also, making assumptions about its size are not very robust. The
updated test checks explicitly for availability of needed classes and
relations.