Skeleton front-end model: emit join event with args in correct order
Winning and losing skeleton got mixed up when factoring the join
functionality in a model function. This caused the neuron name service
to query annotations for the removed neuron, which in turn lead to an
error message after two skeletons were joined. This happened if the
losing skeleton was part of any widget registered to the neuron name
service.  Correct argument order for the event is now restored.