Make tracing overlay guarantee return of selected connector in node query
For treenodes and connectors it makes sense to always also get a
selected node (if any) in addition to the regular spatial query
results. The user usually wants to do something with a selected node, so
its most recent data should be available in the client.

This fixes an issue where new connector links (e.g. postsynaptic
placeholders) are created on a section different from where the
connector is located at.