Volume widget: make mesh update and saving more robust
Meshes are now updated before they are saved, in case they are out of
sync with their properties. In case no mesh could be generated, saving
is canceled and the edit form stays open.

Also, the UI is now blocked until a mesh is re-generated. When saving a
volume, the volume list is only shown once saving was finished. This way
the newly added volume will show up in the list as well.