Navigator tool: update order of default zoom values
The order is expected to be highest to smallest for the zoom level. And
even tough the slider value set is updated for the active stack, the
slider handle position isn't updated even the new slider value equals
the old one. While this should certainly be fixed as well (forcing a
handle update if value set changes), updating the Navigator default
values is enough to get the correct slider position for stacks with
eight zoom levels.

Fixes catmaid/CATMAID#1313