Command history: ignore promise rejections in internal exec/undo
The internal continuation of command execution and undo doesn't need to
take care of errors happening in the actual command. It only needs to be
aware of them to not advance/rollback history in such a situation. To
not have the browser display such an uncaught error in the console, it
is no silently dropped (with CATMAID.noop).