Graph widget: make alt+click connector infor more robust
A connector info request is now only sent if there is request data
available. If the source is a single node and the target a group, no
connector info was shown so far (resulted in silently dropped error).
This is now fixed as well.

Also use CATMAID.fetch() instead of the request queue. This makes the
code more reasable and catches more errors.