Commands: set execution context header in commands
Requests generated during regular command execution, undo or redo will
now have an extra header field: "X-CATMAID-Execution-Context". It can
have either "EXEC", "UNDO" or "REDO" as values, indicating what the
context for execution of the command was.

This information can be used on the back-end to gather statistics on the
use of undo in general, but can also provide important meta data to find
problems.