Add optional New Relic middleware to log execution context
Requests generated through commands will contain a so called request
context header, which indicates if an operation originated in regular
command execution, undo or redo. The new middleware will look for this
header field and will log it to New Relic with key 'execution_context'.
If the header is not present, 'unknown' is used.