Alpha shapes: remove inner edges
The alpha shape module we use tries to remove all inner faces of the
generated mesh, but unfortunately fails. This issue has been reported,
but for now it seems that we can get around this by removing all
duplicate faces in the returned triangle set. Faces that appear more
than once appear to be exactly the inner faces we want to remove.