State: handle state mismatches more gracefully
If a client tries to perform an operation without having the most recent
data available and the performed action is canceled, a more helpful
dialog is now shown. It explains the situation and offers to refresh the
client view (currently only nodes in the tracing layer are refreshed).

This introduces the EVENT_STATE_NEEDS_UPDATE event to the CATMAID.State
namespace. Listeners are requested to refresh their client state if they
receive it.