Add support for PyPy
PyPy is now supported and can be used to improve performance of back-end
heavy endpoints. Most functionality is available, except for the
following: Ontology clustering, Cropping, Synapse clustering, HDF 5
tiles and User analytics. To use PyPy, a new virtualenv using the PyPy
executable has to be created:

mkvirtualenv --no-site-packages -p /usr/bin/pypy catmaid-pypy

The regular requirements.txt file can be used to install dependencies
within this virtualenv. Pip's environment markers are used to specify
which dependencies can be installed with PyPy and which not.