Tracing layer: select joined-in node after skeleton join
This was the case before, but was recently changed due to a mix-up of to/from
nodes in a join in the model back-end. Also wait to activate the joined-in node
until an event-triggered node update finished. This is required for activated
node to have the correct skeleton ID.  This in turn is needed to use the correct
neuron name.

Fixes catmaid/CATMAID#1324