Skeleton source subscriptions: disable pull button by default
The pull button seems to be more confusing than it would help. It does
the same as the "Append" button, but for the selected skeleton
subscription source. Since this is not strictly subscription related and
no widgets without "Append" button exist, a default of not showing the
"Pull" button seems justifiable. It saves some space and reduces some
confusion with the already complex subscription UI.

Also, some whitespace is added if the "Pull" button is not shown. This
helps distinguishing the neighboring checkbox and drop down labels.