Navigator: shorten slice change delay
Reduce the delay when changing the slice with the mouse or keyboard
shortcuts from 100ms to 50ms to make these operations feel more
responsive. On platforms tested this delay was still long enough to
not result in spurious changes when moving the scroll wheel quickly or
holding the key shortcut.

Note that this is a similar change to the one by fzadow in
catmaid/CATMAID#701, but does not change the slider UI component delays.