Tracing layer: suspend hover actions on links if connector is deleted
When a connector is deleted, there is a small time window between
initiating the removal and the removal of the graphic representation of
the connector. If a user hovers over link edges during this window, an
warning is shown (regarding the actually already unavailable connector).
This is now fixed in this commit by suspending mouse actions on links
right before a connector deletion is triggered.

Fixes catmaid/CATMAID#1327