Add history tables
A new migration will install history tables for most CATMAID and
non-CATMAID tables. History tables track all changes in their respective
live tables along with a time range a particular row was valid. Validity
is represented through a half-open time interval (i.e. the beginning is
included in the range, the end is not), called system time period
(sys_period column). Insertion will create a new row not only to the
live table, but also in the history table. The sys_period will be
[creation_time, infinity). Updating a live table row leads to a new
history table entry with a sys_period of [edition_time, infinity). The
previous history entry for the changed object will be updated to be only
valid until this change, i.e. its sys_period becomes [creation_time,
edition_time). Deletion of a live table entry will not create a new
history table entry, but update the corresponding sys_period in the
history table to end with the deletion, i.e. [edition_time,
deletion_time). Tests for this behavior are included.

History entries also record the ID of the current transaction in a
column called exec_transaction_id. Together with the range start a
unique reference is created. Both the time and the transaction ID are
also referenced in another table: catmaid_transaction_info. It allows to
record a user and label for a transaction, which in turns allows CATMAID
now to store semantic information to individiual transactions (similar
to the log table, but referencing history).