Volumes: add interval based alpha shape computation
This code isn't used yet but will replace the current alpha shape
generation soon. It implements basically was has been discussed in [1]
and [2]. For each simplex (and sub-simplex) in the simplicial complex
representing the Delaunay triangulation of the passed in points three
values are stored: its circumradius, a boundary alpha value b and a
interior alpha value i with 0 < b <= i <= infinity. The corresponding
simplex is on the alpha shape boundary iff b <= alpha < i. It is in the
interior of the alph shape iff i <= alpha < infinity. This should allow
much quicker exploration of different alpha values.

[1] H. Edelsbrunner and E. P. Mücke. Three-dimensional alpha shapes. ACM
Trans. Graph., 13(1):43–72, January 1994
[2] K. Fischer. Introduction to Alpha Shapes.