Volume manager: use numeric fields for input for easy changes
Numeric fields can be controlled with the mousewheel and little buttons.
This makes changing alpha values much easier. Whenever the value is
changed, the mesh (and preview) is updated. This is very fast due to the
use of interval based alpha shapes.

A second numeric field allows to adjust the step size of the alpha
input. This allows for flexible exploring the alpha value space.