Alpha shapes: make triangle filtering optional
With interval based alpha shapes, the default was so far to filter
triangles of the simplicial complex representing a 3D triangulation. It
seems, however, that resulting meshes look better and more consistent if
not triangles but tetrahedra are filtered. This has been the case with
non-interval based alpha shapes and causes result meshes to always be
"closed".  With triangle based filtering one often ends up with "open"
shapes, where not all triangles of a hull would satisfy the filtering
criterion. Therefore, filtering tetrahedra is the default again, but
triangle filtering is left as an option.