Alpha shapes: prevent double-updates on alpha changes
The numeric input box for alpha sets the alpha value on a volume based
on multiple events and some of them can happen together (e.g. onmouseup
and onchange). To prevent a double-update of a mesh in these situations,
a mesh is now only refreshed as a consequence of an alpha update, if the
alpha value actually changed.