Skeleton review back-end: improve error message if root not found
If no (relative) root node is found when extracting review information,
an error is now raised. This wasn't the case before and logs indicated
that no root was found at least once recently. To get more information
on the passed in parameters in such cases, this error check is added.