Add basic release script
This script will create a release breanch and prepare CHANGELOG.md,
API_CHANGELOG.md and the Sphic documentation config for a release of
today's date. It wraps these changes in a commit and adds an annotated
tag to this commit. Tag tag message includes the concatenated text of
CHANGELOG.md and API_CHANGELOG.md, with "#" characters replaced by "-".

It does not yet push to a remote and merge the release branch to master
and dev.

See catmaid/CATMAID#1304