Tracing overlay: make node moving/clicking + scrolling more robust
If a node is dragged around or pressed while sections are changed, the
node's mouse handlers could throw an error. This updates the node
identity check to not update the layer in case of a mismatch (because
this won't change the situation) and updates its console log messages.
Also a check is added to the node mouse move handler that ensures the
moved node is available in the tracing layer itself.