Remove catalogueable property of /projects/ endpoint results
The catalogueable property of a project is based on the existence of the
"driver_line" class in a project. It is a semantic aspect of a project
that is dealt with by data views to create links to some other service,
called a catalog, which is respected by some data views. However, for
the /projects/ endpoint, there is no need to treat this ontology
definition for a project special in any way. Many other classes are used
with projects and a generic query interface to filter based on ontology
facts is already available. The /projects/ endpoint should return all
projects visible to user along with their stacks and stack groups.
Other semantic properties should be taken care of separately. This can
be done through the ontology interface.