Neuron name service: move delimiter option into format string
Specifying the delimiter for multi-value components of a neuron name
(e.g. a list of annotations), is now done as part of the format string.
Each %f or %N component can now optionally be followed by "{delimiter}",
where "delimiter" is used to separate values of this component. This
replaces the changes introduced in f4d64e3 with a more consistent
version with less fields to keep track of.

See catmaid/CATMAID#1341