Connectivity matrix: add PDF export (table printing support)
Connectivity matrices can now be exported as PDF. The "Export PDF"
button on the "Main" tab will open a dialog explaining that this PDF
export is made through having the browser print the table and how to
enable colors. Pressing "Print" will show the browser's print dialog for
only the table. All further adjustments are made in the print dialog.

The export looks just like in the browser.

Closes catmaid/CATMAID#1347