Connectors: add general skeleton ID filtering to info query
This allows a "skids" parameter to be passed to the
{project_id}/connector/info endpoint, which will return all
connectors linked in some way to the result connectors. This is useful
to get all connectors for a set of skeletons. This parameter, however,
can't be used together with the pre and post parameters, because they do
their own skeleton based filtering.